requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-writable-nested==0.6.3
gTTS==2.2.4
gunicorn==20.1.0
idna==3.3
numpy==1.22.2
pandas==1.4.1
Pillow==9.1.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
